fix e2e

+++ committer/src/config.rs
@@ -100,7 +100,7 @@ pub struct App {
     #[serde(deserialize_with = "human_readable_duration")]
     pub gas_bump_timeout: Duration,
     /// Max gas fee we permit a tx to have in wei
-    pub tx_max_fee: u128,
+    pub tx_max_fee: u64,
     ///// Contains configs relating to block state posting to l1
     pub bundle: BundleConfig,
     //// Duration for timeout when sending tx requests
